Add a Pharo variant to open a connection with a timeout
In case the destination blackholes (no RST being sent) waiting Stocket standardTimeout is a very long time. Add an overload to use the Pharo sockets directly and allow to specify a timeout.  In the long run I want to have a Context ( https://golang.org/pkg/context) that allows to specify an absolute timeout for a logical operation.