Verify that the insert/update/delete commands were fully carried out
Checking the "success" of the $cmd is not enough to know if the operation was carried out. Raise an exception if there was an error with the writeConcern (e.g. it can't be synced to all nodes) and propagate the last writeError as an Exception (e.g. when a unique key constraint is violated).